test: update playwright onboarding trial flow
