+++ e2e/playwright/tests/smoke.spec.ts
@@ -1,8 +1,11 @@
 import { clerk, clerkSetup } from "@clerk/testing/playwright";
 import { expect, test } from "@playwright/test";
 import { deriveAppUrl, STORAGE_STATE } from "../playwright.config";
+import { fillStripeCheckout } from "../lib/stripe-checkout";
 
 test("sign in and complete onboarding to chat page", async ({ page }) => {
+  test.setTimeout(240_000);
+
   const email = process.env.E2E_CLERK_USER_EMAIL!;
   const appUrl = deriveAppUrl(process.env.VM0_API_URL!);
 
@@ -34,7 +37,7 @@ test("sign in and complete onboarding to chat page", async ({ page }) => {
 
   // Verify: landed on chat page
   await page.waitForURL("**/agents/*/chat", {
-    timeout: 60_000,
+    timeout: 120_000,
     waitUntil: "domcontentloaded",
   });
   expect(page.url()).toMatch(/\/agents\/.*\/chat/);
@@ -66,11 +69,18 @@ async function completeOnboarding(page: import("@playwright/test").Page) {
     await page.getByRole("button", { name: "Next" }).click();
   }
 
-  // Step 2: choose tools. This is the terminal step — "Get Started"
-  // finishes onboarding and navigates into the web chat. The step 1 → step 2
-  // transition runs the eager-init API, so allow plenty of time.
+  // Step 2: choose tools. The step 1 → step 2 transition runs the eager-init
+  // API, so allow plenty of time.
   const chooseTools = page.getByTestId("onboarding-step-select-connectors");
   if (await tryAwaitVisible(chooseTools, 15_000)) {
+    await page.getByRole("button", { name: "Next" }).click();
+  }
+
+  // Step 3: start the Pro trial. Stripe redirects back to onboarding; once the
+  // webhook clears onboardingPaymentPending, the app redirects to the chat page.
+  const trialStep = page.getByTestId("onboarding-step-trial");
+  if (await tryAwaitVisible(trialStep, 15_000)) {
     await page.getByRole("button", { name: "Get Started" }).click();
+    await fillStripeCheckout(page);
   }
 }
